test: add end-to-end suite and cover seven untested services
Adds a Playwright end-to-end suite (22 tests) that drives the real stack —
the real Express API against an ephemeral MongoDB, and the real production
SPA served by the frontend's own Express server — stubbing only outbound
third-party HTTP, so a run needs no credentials and no network. See ADR-017.

Adds the first tests for the seven backend services that had none (user,
plan, role, category, report, admin-analytics, shop-analysis). Backend goes
from 24 files / 387 tests to 33 files / 584 tests.

Between them these found 21 defects. Fixed here:

Security
- A shop owner could create a platform administrator. getAllRoles filtered
  only the role dropdown; registerShopMember and updateMemberRole checked
  that a role existed, not that it was assignable, and roleId is validated
  with isMongoId() alone. The rule is now one exported source of truth
  (NON_ASSIGNABLE_MEMBER_ROLES) enforced at both assignment sites.
- changePassword did not clear Users.refreshToken, so the session a password
  change was meant to evict kept refreshing itself.
- Three more mass-assignment holes: updateCategory (move a category onto a
  competitor's public menu), updatePlan (repoint a live Paymob mandate) and
  updateRole. Allowlists, each tested in both directions.
- confirmEmailChange re-checks the destination address, so a collision is a
  400 rather than a raw driver error surfacing as a 500.

Fixed, found by the E2E suite
- The order cancel button rendered the literal string "cancelOrder" in both
  languages; neither it nor "cancelling" existed in either locale file.
- The language reset to English on every page load — nothing was persisted,
  so a refresh, a shared link or a scanned QR code lost it, RTL included.
- No dish could be created from the dashboard at all: the image schema
  rejected the form's own empty value, and ItemImageSection overwrote
  react-hook-form's onChange so the FileList never reached it.

Fixed, other
- getTopPerformingRestaurants returned [] whenever dates were omitted, an
  Invalid Date in $match matching nothing. Bad dates now 400.
- The README embedded docs/screenshots/07.png, deleted in 0347d92, so the
  repository front page rendered a broken image. Renumbered to 17.

Also: .cache/ and e2e/.artifacts/ are now gitignored (a 77 MB mongod binary
and per-run videos), and MONGOMS_DOWNLOAD_DIR is set for local backend runs
so local, CI and E2E share one cached binary.

+++ backend/validators/shop.validator.ts
@@ -68,14 +68,20 @@ export const updateShopValidator = [
 ];
 
 export const validateRegenerateQRCode = [
+  // .toInt() matters as much as the bounds here: isInt() accepts the *string*
+  // "600" as valid, and the handler passes the value straight to QRCode as a
+  // pixel width. Without the sanitiser a well-formed request can still reach
+  // the encoder with a string where it expects a number.
   body("width")
     .optional()
     .isInt({ min: 100, max: 1000 })
-    .withMessage("Width must be between 100 and 1000"),
+    .withMessage("Width must be between 100 and 1000")
+    .toInt(),
   body("margin")
     .optional()
     .isInt({ min: 0, max: 10 })
-    .withMessage("Margin must be between 0 and 10"),
+    .withMessage("Margin must be between 0 and 10")
+    .toInt(),
   body("errorCorrectionLevel")
     .optional()
     .isIn(["L", "M", "Q", "H"])
